위경도조회시 소수추가

+++ BackEnd/demo/src/main/java/com/example/demo/dao/LikeSpotDao.java
@@ -26,9 +26,11 @@ public class LikeSpotDao {
         repository.deleteById(idx);
     }
 
-    public List<LikeSpot> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Integer smallLa, Integer largeLa, Integer smallLo,
-            Integer largeLo, String memberIdx) {
-        return repository.findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(smallLa, largeLa, smallLo, largeLo, memberIdx);
+    public List<LikeSpot> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Double smallLa, Double largeLa,
+            Double smallLo,
+            Double largeLo, String memberIdx) {
+        return repository.findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(smallLa, largeLa, smallLo, largeLo,
+                memberIdx);
     }
 
 }
