+++ BackEnd/demo/src/main/java/com/example/demo/repository/LikeSpotRepository.java
@@ -6,14 +6,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.dto.LikeSpot;
 
-public interface LikeSpotRepository extends JpaRepository<LikeSpot, Long>{
+public interface LikeSpotRepository extends JpaRepository<LikeSpot, Long> {
 
     List<LikeSpot> findByMemberIdx(String id);
 
     void deleteById(Long idx);
 
-    List<LikeSpot> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Integer smallLa, Integer largeLa, Integer smallLo,
-            Integer largeLo, String memberIdx);
-            
+    List<LikeSpot> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Double smallLa, Double largeLa, Double smallLo,
+            Double largeLo, String memberIdx);
 
 }
