+++ BackEnd/demo/src/main/java/com/example/demo/repository/ReviewRepository.java
@@ -7,8 +7,8 @@ import com.example.demo.dto.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 
-    List<Review> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Integer smallLa, Integer largeLa, Integer smallLo,
-            Integer largeLo, String memberIdx);
+    List<Review> findByLatitudeBetweenAndLongitudeBetweenAndMemberIdx(Double smallLa, Double largeLa, Double smallLo,
+            Double largeLo, String memberIdx);
 
     List<Review> findBymemberIdx(String memberIdx);
